Crate async_codec_lite

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesCodec

BytesMut

A unique reference to a contiguous slice of memory.

CborCodec

Framed

FramedParts

FramedRead

FramedWrite

JsonCodec

LengthCodec

LimitCodec

LinesCodec

OverflowError

Enums

LimitError

Traits

Decoder

DecoderWithSkipAhead

Encoder

SkipAheadHandler